requirements.txt
crawl4ai
beautifulsoup4
requests
scikit-learn
langchain-community
langchain
langchain-openai
python-dotenv
faiss-cpu